Add rule in automemeber with focus
In the form, there is Select which doesnt have the input type, so the button type=submit doesnt work with enter there. So I implemented he focus. Even better would be for the PF to focus first OUR clickable element, so it would focus on the add button automatically. We need to do the button type submit nevertheless, this is just for me to store the version with focus to be deleted